Concentrated hash and eq inside TrelloBase to avoid repetition
