+++ trello/attachments.py
@@ -1,12 +1,15 @@
 # -*- coding: utf-8 -*-
 from dateutil import parser as dateparser
 
+from trello import TrelloBase
 
-class Attachments(object):
+
+class Attachments(TrelloBase):
     """
     https://developers.trello.com/advanced-reference/card#get-1-cards-card-id-or-shortlink-attachments
     """
     def __init__(self, id, bytes, date, edge_color, idMember, is_upload, mime_type, name, previews, url):
+        super().__init__()
         self.id = id
         self.bytes = bytes
         self.date = dateparser.parse(date)
@@ -35,15 +38,6 @@ class Attachments(object):
 
     def __repr__(self):
         return u"<Attachments {0}>".format(self.name)
-
-    def __hash__(self):
-        class_name = type(self).__name__
-        return hash(class_name) ^ hash(self.id)
-
-    def __eq__(self, other):
-        if isinstance(other, type(self)):
-            return hash(self) == hash(other)
-        raise NotImplementedError
 
 
 class AttachmentsPreview(object):
