+++ trello/label.py
@@ -1,13 +1,16 @@
 # -*- coding: utf-8 -*-
 from __future__ import with_statement, print_function, absolute_import
+
+from trello import TrelloBase
 from trello.compat import force_str
 
 
-class Label(object):
+class Label(TrelloBase):
     """
     Class representing a Trello Label.
     """
     def __init__(self, client, label_id, name, color=""):
+        super().__init__()
         self.client = client
         self.id = label_id
         self.name = name
@@ -33,15 +36,6 @@ class Label(object):
 
     def __repr__(self):
         return force_str(u'<Label %s>' % self.name)
-
-    def __hash__(self):
-        class_name = type(self).__name__
-        return hash(class_name) ^ hash(self.id)
-
-    def __eq__(self, other):
-        if isinstance(other, type(self)):
-            return hash(self) == hash(other)
-        raise NotImplementedError
 
     def fetch(self):
         """Fetch all attributes for this label"""
