+++ trello/member.py
@@ -1,29 +1,23 @@
 # -*- coding: utf-8 -*-
 from __future__ import with_statement, print_function, absolute_import
+
+from trello import TrelloBase
 from trello.compat import force_str
 
 
-class Member(object):
+class Member(TrelloBase):
     """
     Class representing a Trello member.
     """
 
     def __init__(self, client, member_id, full_name=''):
+        super().__init__()
         self.client = client
         self.id = member_id
         self.full_name = full_name
 
     def __repr__(self):
         return force_str(u'<Member %s>' % self.id)
-
-    def __hash__(self):
-        class_name = type(self).__name__
-        return hash(class_name) ^ hash(self.id)
-
-    def __eq__(self, other):
-        if isinstance(other, type(self)):
-            return hash(self) == hash(other)
-        raise NotImplementedError
 
     def fetch(self):
         """Fetch all attributes for this member"""
